feat(static memory planning): add isDag function to graphZant
Static memory planning is easier if done on an acyclic graph, for we
don't have to take into account outputs that go back to the input of a
prior node, which make the lifetime of such tensors a bit more complex
to model.

The first draft of the static memory planning system won't deal with
cycles, so it might not be suitable for recurrent NNs. This can be
expanded on later on to include support for those if needed.

As a sanity check to not end up in infinite loops or output inconsistent
result, we first assert that the graph is a DAG